requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
PyYAML>=6.0
matplotlib>=3.5.0